+++ ui/packages/shared/client/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.16.105 (2024-02-22)
+
+**Note:** Version bump only for package @parca/client
+
 ## 0.16.104 (2024-02-22)
 
 **Note:** Version bump only for package @parca/client