+++ ui/packages/shared/hooks/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.0.43 (2024-02-22)
+
+**Note:** Version bump only for package @parca/hooks
+
 ## 0.0.42 (2024-02-22)
 
 **Note:** Version bump only for package @parca/hooks