+++ ui/packages/shared/store/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.16.131 (2024-02-22)
+
+**Note:** Version bump only for package @parca/store
+
 ## 0.16.130 (2024-02-22)
 
 **Note:** Version bump only for package @parca/store